% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseresult.R
\name{summary_result}
\alias{summary_result}
\alias{summary_result_dataset}
\title{Parse workflow results}
\usage{
summary_result_dataset(inputTribble)

summary_result(submissionId, workflowId = NULL)
}
\arguments{
\item{inputTribble}{submission IDs (and workflow IDs) of terra job and 
corresponding name}

\item{submissionId}{character(1) submission ID of terra job}

\item{workflowId}{character(1) workflow ID within the submission ID of 
terra job if multiple workflows in one submission}
}
\description{
`summary_result_dataset()` returns csv files of parsing 
  workflow results for different datasets of each workflow steps

`summary_result()` returns csv files of parsing workflow 
  results for different autoBlockSize as well as cpu
}
\examples{
library(AnVIL)
library(Rcollectl)
avworkspace("bioconductor-rpci-yubo/Rcollectlworkflowh5ad")

# inputTribble contains only submission IDs
inputTribble <- tribble(
  ~name, ~submissionId,
  "dgCMatrix_medium", "17efa80e-3ef1-429d-8359-58c7cd653310",
  "DA_medium", "2f0cc8ee-ae53-44e6-a6c0-31e7089f0d42",
  "dgCMatrix_large", "c3c147eb-2ff8-4cf0-ba55-5708fa9c0ae5",
  "DA_large", "8be18771-f06d-4256-ac13-6566a3053f14"
)

# inputTribble contains submission IDs and workflow IDs
inputTribble <- tribble(
  ~name, ~submissionId, ~workflowId,
  "dgCMatrix_medium", "123895c9-6ad1-4258-b3d4-526e057c5cba", "923c360d-a30d-4013-b9f3-80058be1b34a",
  "DA_medium", "123895c9-6ad1-4258-b3d4-526e057c5cba", "c7d5a05a-a275-411d-9908-3e344200cd4e",
  "dgCMatrix_large", "9385fd75-4cb7-470f-9e07-1979e2c8f193", "31c69e74-12d4-40e7-b593-0f59006aca37",
  "DA_large", "9385fd75-4cb7-470f-9e07-1979e2c8f193", "1410f324-021a-43f0-90bd-78fbb6a25c50"
)

summary_result_dataset(inputTribble)
submissionId <- "8ca42a70-d2e9-421f-9f51-f42601348a82"
summary_result(submissionId)
}
