% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{update_data_table}
\alias{update_data_table}
\alias{get_manifest_tibble}
\alias{update_file_avtable}
\title{Update Terra data table}
\usage{
get_manifest_tibble()

update_file_avtable(manifest_tibble, id, sample, dgCMatrix, core, mem_gb)
}
\arguments{
\item{manifest_tibble}{character() a tibble contains result of cellxgene 
data}

\item{id}{character(1) unique id of file table}

\item{sample}{character(1) column to analyze}

\item{dgCMatrix}{logical(1) If TRUE, read in assay as in-memory `dgCMatrix`}

\item{core}{integer() number of cpus}

\item{mem_gb}{numeric() memory (GB) to allocate for running the workflow, 
default: 60}
}
\value{
`get_manifest_tibble()` returns a tibble
}
\description{
`get_manifest_tibble()` retrieve cellxgene data

`update_file_avtable()` imports cellxgene data to Terra data 
  table file
}
\examples{
library(cellxgenedp)
manifest_tibble <- get_manifest_tibble() |>
  filter(
  title == "Immunophenotyping of COVID-19 and influenza highlights the role 
    of type I interferons in development of severe COVID-19",
  filetype == "H5AD"
)
library(AnVIL)
avworkspace("bioconductor-rpci-yubo/Rcollectlworkflowh5ad")
manifest_tibble <- get_manifest_tibble() |>
  filter(
  title == "Immunophenotyping of COVID-19 and influenza highlights the role 
    of type I interferons in development of severe COVID-19",
  filetype == "H5AD"
)
sample = "Disease.group"
id = sprintf("\%04d", dim(avtable("file"))[1] + 1)
update_file_avtable(manifest_tibble, id, sample, TRUE, 3, 60)
}
